kns_indexer bugfix, refactor eth again
